+++ types/coin_test.go
@@ -563,46 +563,18 @@ func (s *coinTestSuite) TestAddCoins() {
 		expected sdk.Coins
 		msg      string
 	}{
-		{"adding two empty lists", s.emptyCoins, s.emptyCoins, s.emptyCoins, "empty, non list should be returned"},
-		{"empty list + set", s.emptyCoins, cA0M1, sdk.Coins{s.cm1}, "zero coins should be removed"},
-		{"empty list + set", s.emptyCoins, cA1M1, cA1M1, "zero + a_non_zero = a_non_zero"},
-		{"set + empty list", cA0M1, s.emptyCoins, sdk.Coins{s.cm1}, "zero coins should be removed"},
-		{"set + empty list", cA1M1, s.emptyCoins, cA1M1, "a_non_zero + zero  = a_non_zero"},
-		{
-			"{1atom,1muon}+{1atom,1muon}", cA1M1, cA1M1,
-			sdk.Coins{s.ca2, s.cm2},
-			"a + a = 2a",
-		},
-		{
-			"{0atom,1muon}+{0atom,0muon}", cA0M1, cA0M0,
-			sdk.Coins{s.cm1},
-			"zero coins should be removed",
-		},
-		{
-			"{2atom}+{0muon}",
-			sdk.Coins{s.ca2},
-			sdk.Coins{s.cm0},
-			sdk.Coins{s.ca2},
-			"zero coins should be removed",
-		},
-		{
-			"{1atom}+{1atom,2muon}",
-			sdk.Coins{s.ca1},
-			sdk.Coins{s.ca1, s.cm2},
-			sdk.Coins{s.ca2, s.cm2},
-			"should be correctly added",
-		},
-		{
-			"{0atom,0muon}+{0atom,0muon}", cA0M0, cA0M0, s.emptyCoins,
-			"sets with zero coins should return empty set",
-		},
+		{"{1atom,1muon}+{1atom,1muon}", sdk.Coins{s.ca1, s.cm1}, sdk.Coins{s.ca1, s.cm1}, sdk.Coins{s.ca2, s.cm2}},
+		{"{0atom,1muon}+{0atom,0muon}", sdk.Coins{s.ca0, s.cm1}, sdk.Coins{s.ca0, s.cm0}, sdk.Coins{s.cm1}},
+		{"{2atom}+{0muon}", sdk.Coins{s.ca2}, sdk.Coins{s.cm0}, sdk.Coins{s.ca2}},
+		{"{1atom}+{1atom,2muon}", sdk.Coins{s.ca1}, sdk.Coins{s.ca1, s.cm2}, sdk.Coins{s.ca2, s.cm2}},
+		{"{0atom,0muon}+{0atom,0muon}", sdk.Coins{s.ca0, s.cm0}, sdk.Coins{s.ca0, s.cm0}, sdk.Coins(nil)},
 	}
 
 	for _, tc := range cases {
 		s.T().Run(tc.name, func(t *testing.T) {
 			res := tc.inputOne.Add(tc.inputTwo...)
 			require.True(t, res.IsValid(), fmt.Sprintf("%s + %s = %s", tc.inputOne, tc.inputTwo, res))
-			require.Equal(t, tc.expected, res, tc.msg)
+			require.Equal(t, tc.expected, res, "sum of coins is incorrect")
 		})
 	}
 }
@@ -611,13 +583,13 @@ func (s *coinTestSuite) TestAddCoins() {
 // are correctly coalesced. Please see issue https://github.com/cosmos/cosmos-sdk/issues/13234
 func TestCoinsAddCoalescesDuplicateDenominations(t *testing.T) {
 	A := sdk.Coins{
-		{"den", math.NewInt(2)},
-		{"den", math.NewInt(3)},
+		{"den", sdk.NewInt(2)},
+		{"den", sdk.NewInt(3)},
 	}
 	B := sdk.Coins{
-		{"den", math.NewInt(3)},
-		{"den", math.NewInt(2)},
-		{"den", math.NewInt(1)},
+		{"den", sdk.NewInt(3)},
+		{"den", sdk.NewInt(2)},
+		{"den", sdk.NewInt(1)},
 	}
 
 	A = A.Sort()
@@ -625,10 +597,10 @@ func TestCoinsAddCoalescesDuplicateDenominations(t *testing.T) {
 	got := A.Add(B...)
 
 	want := sdk.Coins{
-		{"den", math.NewInt(11)},
+		{"den", sdk.NewInt(11)},
 	}
 
-	if !got.Equal(want) {
+	if !got.IsEqual(want) {
 		t.Fatalf("Wrong result\n\tGot:   %s\n\tWant: %s", got, want)
 	}
 }